requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
django-filter==23.2
django-heroku==0.3.1
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-framework-jwt-oauth2==0.0.4
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-registration==0.1.3
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.2
flake8==6.1.0
flake8-docstrings==1.7.0
gunicorn==21.2.0
identify==2.5.26
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
phonenumbers==8.13.17
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.3.3
pre-commit-hooks==4.4.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.1.0
PyJWT==1.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
types-cryptography==3.3.23.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
xlwt==1.3.0
django-redis==5.3.0
